requirements.txt
google-generativeai
openai
pillow
streamlit
python-dotenv
requests